Fix tests in api calls
